+++ HISTORY.md
@@ -15,6 +15,9 @@ Work in progress
     - fixed handling of multi-language sed files
     - extended documentation
     - fixed minor problems
+  - package glossaries
+    - update for new `.glsdef` format introduced in glossaries 4.47; **draft thanks to @mstmob** (PR [#238](../../pull/238), PR [#220](../../pull/220), issue [#219](../../issues/219))
+    - added macros \\acrpluralsuffix, \\glspluralsuffix, \\glsacrpluralsuffix, \\glsupacrpluralsuffix
 
 Version 1.3.1 (2022/08/04)
 --------------------------